+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Added
 - Add support non-mirrored local Swift Packages in `localPackages`. [#796](https://github.com/yonaskolb/XcodeGen/pull/796) @freddi-kit
+- Add buildImplicitDependencies as an option on TargetScheme. [#810](https://github.com/yonaskolb/XcodeGen/pull/810) @evandcoleman
 
 #### Fixed
 - Fixed resolving path to local Swift Packages [#796](https://github.com/yonaskolb/XcodeGen/pull/796) @freddi-kit